query results
Add ability to run a query and get results. Because query results are
async, we need to loop to check until they are ready. I took a gues at
some wait parameters, and it seems to work, but open to suggestions.

No tests written because I don't have a test honeycomb account I can use
with this.